Apply the full memberlist data for map markers
