+++ src/app/components/Map/LocatorMap.tsx
@@ -1,13 +1,11 @@
 import { MapContainer, TileLayer } from 'react-leaflet';
 import { useAppSelector } from '../../redux/hooks';
-import { selectMemberShortList } from '../../redux/slices/memberShortListSlice';
 import MemberMapMarker from './MemberLocationMarker';
 import MyLocationMarker from './MyLocationMarker';
-import { selectSortedMemberList } from '@/app/redux/slices/memberListSlice';
+import { selectMemberList } from '@/app/redux/slices/memberListSlice';
 
-const Map = () => {
-  const { list: memberShortList, isLoading, hasError } = useAppSelector(selectMemberShortList);
-  const list = useAppSelector(selectSortedMemberList);
+const LocationMap = () => {
+  const memberList = useAppSelector(selectMemberList);
 
   return (
     <div className="h-80 lg:h-96 w-full">
@@ -20,7 +18,7 @@ const Map = () => {
         <TileLayer
           url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
         />
-        {!isLoading && memberShortList.map((memberInfo) => (
+        {memberList.map((memberInfo) => (
           <MemberMapMarker
             key={memberInfo.id}
             markerData={{
@@ -36,6 +34,6 @@ const Map = () => {
   )
 }
 
-export default Map;
+export default LocationMap;
