+++ src/app/components/MemberList/MemberGridList.tsx
@@ -1,18 +1,17 @@
-import React, { useEffect } from 'react';
+import React, { useMemo } from 'react';
 import { useAppSelector } from '@/app/redux/hooks';
-import { selectSortedMemberList } from '@/app/redux/slices/memberListSlice';
+import { selectMemberList } from '@/app/redux/slices/memberListSlice';
 import { selectMemberShortList } from '@/app/redux/slices/memberShortListSlice';
-import { MemberInfo } from '@/app/redux/types';
 import MemberCard from './MemberCard';
 
 const MemberGridList = () => {
   const { list: memberShortList } = useAppSelector(selectMemberShortList);
-  const sortedList = useAppSelector(selectSortedMemberList);
+  const memberList = useAppSelector(selectMemberList);
 
-  const renderList = sortedList.length > 0 
-    ? sortedList : memberShortList.map((member) => ({...member, distance: undefined}));
+  const renderList = useMemo(() => 
+    memberList.length > 0 ? memberList : memberShortList.map((member) => ({...member, distance: undefined}))
+    ,[memberList, memberShortList]);
   
-  console.log({renderList});
   return (
     <div className="h-80 lg:h-96 overflow-y-scroll my-4 px-2">
       <div className="grid grid-cols-2 gap-4 xl:grid-cols-3 py-4 px-2">
