+++ src/app/redux/slices/userEventDataSlice.ts
@@ -3,10 +3,10 @@ import type { RootState } from '../types';
 import { DetailedInfo, Location } from '../types';
 
 const initialState= {
-  myLocation: {
-    lat: 0,
-    long: 0,
-  } as Location,
+  // myLocation: {
+  //   lat: 0,
+  //   long: 0,
+  // } as Location,
   highlightedMemberId: '',
   isModalOpen: false,
   selectedMemberDetails: {} as DetailedInfo,
@@ -16,9 +16,9 @@ export const userEventDataSlice = createSlice({
   name: 'userEventSlice',
   initialState,
   reducers: {
-    setLocation: (state, action) => {
-      state.myLocation = action.payload;
-    },
+    // setLocation: (state, action) => {
+    //   state.myLocation = action.payload;
+    // },
     setHighlightedMemberId: (state, action) => {
       state.highlightedMemberId = action.payload;
     },
@@ -33,11 +33,11 @@ export const userEventDataSlice = createSlice({
 
 export const {
   setHighlightedMemberId,
-  setLocation,
+  // setLocation,
   setIsModalOpen,
   setSelectedMemberDetails,
 } = userEventDataSlice.actions;
-export const selectMyLocation = (state: RootState) => state.userEventData.myLocation;
+// export const selectMyLocation = (state: RootState) => state.userEventData.myLocation;
 export const selectHighlightedMemberId = (state: RootState) => state.userEventData.highlightedMemberId;
 export const selectIsModalOpen = (state: RootState) => state.userEventData.isModalOpen;
 export const selectSelectedMemberDetails = (state: RootState) => state.userEventData.selectedMemberDetails;
