+++ src/app/redux/types.ts
@@ -11,15 +11,40 @@ export type MemberShortInfo = {
   long: number;
 }
 
-export type MemberShortInfoForSorting = {
-  distance?: number;
-} & MemberShortInfo;
+// export type MemberExtraInfo = {
+//   id: string;
+//   name?: string;
+//   height?: string;
+//   mass?: string;
+//   gender?: string;
+//   homeworld?: string;
+//   wiki?: string;
+//   image?: string;
+//   born?: string;
+//   died?: string;
+//   diedLocation?: string;
+//   species?: string;
+//   hairColor?: string;
+//   eyeColor?: string;
+//   skinColor?: string;
+//   cybernetics?: string;
+//   affilications?: string[];
+//   masters?: string[];
+//   apprentices?: string[];
+//   formerAffilications?: string[];
+// }
 
 export type DetailedInfo = {
   [key: string]: string | string[] | undefined | number;
 }
 
-export type MemberInfo = MemberShortInfoForSorting & DetailedInfo;
+export type MemberInfo = MemberShortInfo & DetailedInfo & {
+  distance?: number;
+};
+
+export type MemberWithDistance = MemberInfo & {distance: number};
+
+export type MemberListWithDistance = MemberWithDistance[];
 
 export type Location = {
   lat: number;
